Changed how screen iterators are handled when using the safe interface. XCBScreenIterator now no longer is a wrapper struct; It simply is a type alias to cdef::XCBScreenIterator now. The added indirection was redundant.
Added a function xcb_screen to access an XCBScreenIterator's current screen. This was done mainly because the safe interface tries to keep application code free from unsafe blocks that are the result of dereferencing raw pointers.

The function xcb_screen_next has been updated to reflect the changes to XCBScreenIterator.

Implemented dereferencing for the DoNotFree wrapper struct in order to be able to access a screen's fields in application code. This should also come in handy for other resources.

The simple_window example has been updated accordingly.